Add basic authentication system
Clients can now login with the LOGIN command. This is only required when
a users.txt file exists.

The command format is "LOGIN <username> <password>"
Passwords should be base64 encoded (yes, I know).

The storage format is one user per line, with the username, a space, and
then an MD5 hash (yes, I also know) in hex form (non-binary) of the
decoded password. There is no mechanism for creating new accounts. They
must be added manually.

I'm aware that this is not best security practices. This software is not
meant to be openly accessible to the public. The system may be improved
upon in the future if deemed necessary.