Allow clients to resume connections
Remote connections are no longer closed on client disconnect. Clients
can use the RESUME command to relink with an old connection and begin
receiving data from it again.

A client can only RESUME a connection that it created, unless the
account system is disabled, in which case they can resume any
connection.